requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
bitarray==3.8.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
ckzg==2.1.5
construct==2.10.70
construct-typing==0.7.0
cytoolz==1.1.0
Django==5.2.9
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonalias==0.1.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
multidict==6.7.0
packaging==25.0
parsimonious==0.10.0
pillow==12.0.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
pytest==9.0.2
pytest-django==4.11.1
pytest-mock==3.15.1
pyunormalize==17.0.0
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.7
requests==2.32.5
rlp==4.1.0
rpds-py==0.30.0
rsa==4.9.1
solana==0.36.10
solders==0.27.1
sqlparse==0.5.5
toolz==1.1.0
tqdm==4.67.1
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
web3==7.14.0
websockets==15.0.1
yarl==1.22.0
whitenoise==6.7.0
